+++ CHANGELOG.md
@@ -70,7 +70,7 @@
 
 ## [2.0.9] - 2017-11-09
 
-	* Fix issue 77. Template filter xtd_comment_gravatar_url must not
+	* Fix issue 77. Template filter tree_comment_gravatar_url must not
 	  hard-code http schema in URL (reported by @pamost).
 
 ## [2.0.8] - 2017-09-24